requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
ndeflib==0.3.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyperclip==1.11.0
PyRect==0.2.0
pyscard==2.3.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pyusb==1.3.1
